proxy ip/port validation
now properly retrieves the client IP and port, given a secret auth token is provided by Cloudflare.
this required setting up two header transform rules, a dynamic rule for the client port, and a static rule for the authentication. theoretically, it is possible to connect to the server without going through Cloudflare, so this auth was added for validation. I should really see about getting a firewall set on the host.