change ip token tracking
due to the nature of how HTTP requests are made, controlling the port used to send information from is too difficult (or perhaps impossible). IPs can only be whitelisted, ports cannot be a part of that. sadge.

But does this degrade security? Maybe a little... but ultimately if the exchange is leaked then that's a big F on it's own, nothing could stop the potential DOS that would cause anyway.