Change how logout times are handled
In the process of a major refactor for the main server plugins. For now, lastLogout will always contain the most recent logout time. In the future, a new value will be used to track successful login/ logout cycles. Also, keeping the lastLogout time is better than changing existing data to track something not immediately obvious. It is far more clear that a lastLogout time which is before the lastLogin time means the player is currently connected. Even so, I would rather return a specific value that literally says "yes, this player is connected to the server right now" instead of forcing callers to do a math operation to check this.